backend/requirements.txt
django>=5.0,<5.1
django-environ
djangorestframework
psycopg2-binary
django-cors-headers
cryptography
